requirements.txt
asgiref==3.11.1
Django==6.0.2
django-phonenumber-field==8.4.0
djangorestframework==3.16.1
dotenv==0.9.9
phonenumbers==9.0.25
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
